+++ test/loader.js
@@ -13,16 +13,43 @@ const loader = require( '..' );
 
 describe( 'freemarker-loader', () => {
 
-   it( 'runs in the pitching phase', done => {
-      pitchLoader( {
-         query: '?data=data.js&classpath=freemarker.jar'
-      }, __dirname + '/index.html', done );
+   it( 'runs', done => {
+      runLoader( {
+         query: '?template=' + __dirname + '/index.html&classpath=freemarker.jar',
+         callback( err, source ) {
+            if( err ) {
+               done( err );
+               return;
+            }
+
+            expect( dropEmptyLines( source ) ).to.equal( [
+               '<h1>Date: Nov 12, 1955 10:04:00 PM</h1>',
+               '<dl>',
+               '   <dt><em>twelve</em>:</dt><dd>12</dd>',
+               '   <dt><em>fourty three</em>:</dt><dd>43</dd>',
+               '   <dt><em>one hundred and three</em>:</dt><dd>123</dd>',
+               '</dl>'
+            ].join( '\n' ) );
+            done();
+         }
+      }, __dirname + '/data.js' );
    } );
 
-   function pitchLoader( options, request, done ) {
-      loader.pitch.call( Object.assign( {
+   function dropEmptyLines( string ) {
+      return string
+         .split( '\n' )
+         .filter( line => !(/^\s*$/.test( line )) )
+         .join( '\n' );
+   }
+
+   function runLoader( options, request ) {
+      const resourcePath = request.split( '!' ).pop();
+      const source = fs.readFileSync( resourcePath );
+
+      loader.call( Object.assign( {
+         options: { context: __dirname },
          context: __dirname,
-         resourcePath: request.split( '!' ).pop(),
+         resourcePath,
          fs,
          cacheable() {},
          async() {},
@@ -32,17 +59,15 @@ describe( 'freemarker-loader', () => {
             fs.readFile( filename, callback );
          },
          exec( source ) {
+            // eslint-disable-next-line no-new-func
             const fn = new Function( 'module', 'exports', source );
             const module = {
                exports: {}
             };
             fn( module, module.exports );
             return module.exports;
-         },
-         callback( err, source ) {
-            done( err );
          }
-      }, options ), request, '', {} );
+      }, options ), source, {} );
    }
 
 });